graph/data/store, store/postgres: Move ToSql impl to store/postgres
Move dependencies (BigDecimal, Diesel) out of graph/data/store.
Create new struct SqlValue(Value) in Store/Postgres and implement
ToSql trait on it.
Use SqlValue in filter queries
