+++ graph/src/lib.rs
@@ -1,6 +1,3 @@
-extern crate bigdecimal;
-#[macro_use]
-extern crate diesel;
 extern crate ethabi;
 extern crate ethereum_types;
 extern crate futures;
